+++ MiamIOSFramework/MiamIOSFramework/ui/RecipeDetails/RecipeDetailsDifficulty.swift
@@ -22,26 +22,26 @@ struct RecipeDetailsDifficulty: View {
             switch difficulty {
             case 1 :
                 HStack{
-                    Image.miamImage(icon: .easy).frame( height: 24.0, alignment: .center)
+                    Image.miamImage(icon: .difficultyLow).frame( height: 24.0, alignment: .center)
                    
                 }
                 Text(MiamText.sharedInstance.difficultyEasy)
                     .foregroundColor(Color.miamColor(.secondaryText)).font(.system(size: 13.0, weight: .regular, design: .default))
             case 2 :
                 HStack{
-                    Image.miamImage(icon: .mid).frame( height: 24.0, alignment: .center)
+                    Image.miamImage(icon: .difficultyMedium).frame( height: 24.0, alignment: .center)
                 }
                 Text(MiamText.sharedInstance.difficultyMid)
                     .foregroundColor(Color.miamColor(.secondaryText)).font(.system(size: 13.0, weight: .regular, design: .default))
             case 3 :
                 HStack{
-                    Image.miamImage(icon: .hard).frame(width: 24.0, height: 24.0, alignment: .center)
+                    Image.miamImage(icon: .difficultyHigh).frame(width: 24.0, height: 24.0, alignment: .center)
                 }
                 Text(MiamText.sharedInstance.difficultyHard)
                     .foregroundColor(Color.miamColor(.secondaryText)).font(.system(size: 13.0, weight: .regular, design: .default))
             default :
                 HStack{
-                    Image.miamImage(icon: .easy).frame( height: 24.0, alignment: .center)
+                    Image.miamImage(icon: .difficultyLow).frame( height: 24.0, alignment: .center)
                    
                 }
                 Text(MiamText.sharedInstance.difficultyEasy)