+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/theme/template.kt
@@ -215,6 +215,9 @@ object Template {
      */
     var CatalogResultPageLazyLoaderTemplate: (@Composable() () -> Unit)? = null
 
+    var CatalogPageTitleTemplate: (@Composable() (
+        title: String
+    ) -> Unit)? = null
 
     var CatalogFavoritEmptyTemplate: (@Composable() (
         returnToCatalog: () -> Unit
@@ -248,7 +251,7 @@ object Template {
     var EquipmentPreferencesSectionTemplate: (@Composable() (
         equipmentsTag: List<CheckableTag>, togglePreference: (tagIdToToogle: String) -> Unit
     ) -> Unit)? = null
-    
+
 ////////////////// MY MEAL BUTTON ////////////
 
     var myMealButtonSuccessViewTemplate: (@Composable() (