+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/catalog/CatalogPage.kt
@@ -1,6 +1,7 @@
 package com.miam.kmm_miam_sdk.android.ui.components.catalog
 
 import android.content.Context
+import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -10,7 +11,9 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.GridCells
+import androidx.compose.foundation.lazy.GridItemSpan
+import androidx.compose.foundation.lazy.LazyVerticalGrid
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.CircularProgressIndicator
@@ -42,6 +45,9 @@ import com.miam.kmm_miam_sdk.android.ui.components.states.ManagementResourceStat
 fun CatalogPage(
     recipePageVM: RecipeListPageViewModel,
     context: Context,
+    columns: Int,
+    verticalSpacing: Int,
+    horizontalSpacing: Int,
     returnToCategoriesPage: () -> Unit
 ) {
 
@@ -54,6 +60,9 @@ fun CatalogPage(
             CatalogSuccessPage(
                 recipePageVM,
                 recipes,
+                columns,
+                verticalSpacing,
+                horizontalSpacing,
                 context
             )
         },
@@ -74,22 +83,36 @@ fun CatalogPage(
     )
 }
 
+@OptIn(ExperimentalFoundationApi::class)
 @Composable
 private fun CatalogSuccessPage(
     recipePageVM: RecipeListPageViewModel,
     recipes: List<Recipe>,
+    columns: Int,
+    verticalSpacing: Int,
+    horizontalSpacing: Int,
     context: Context
 ) {
-        LazyColumn(
+    Box(modifier = Modifier.fillMaxSize()) {
+        LazyVerticalGrid(
             modifier = FavoritePageStyle.favoriteMainContainer,
+            cells = GridCells.Fixed(columns),
+            verticalArrangement = Arrangement.spacedBy(verticalSpacing.dp, Alignment.Top),
+            horizontalArrangement = Arrangement.spacedBy(horizontalSpacing.dp, Alignment.Start)
         ) {
-            item {
-                Row {
-                    Text(
-                        text = recipePageVM.currentState.title,
-                        color = Colors.black,
-                        style = Typography.subtitleBold
-                    )
+            item(span = { GridItemSpan(columns) }) {
+                if (Template.CatalogPageTitleTemplate != null) {
+                    Template.CatalogPageTitleTemplate?.let {
+                        it(recipePageVM.currentState.title)
+                    }
+                } else {
+                    Row {
+                        Text(
+                            text = recipePageVM.currentState.title,
+                            color = Colors.black,
+                            style = Typography.subtitleBold
+                        )
+                    }
                 }
             }
             itemsIndexed(recipes) { index, item ->
@@ -99,21 +122,23 @@ private fun CatalogSuccessPage(
                 recipe.Content()
                 if (index == recipes.lastIndex) {
                     recipePageVM.setEvent(RecipeListPageContract.Event.LoadPage)
-                    if (recipePageVM.currentState.isFetchingNewPage) {
-                        if (Template.CatalogResultPageLazyLoaderTemplate != null) {
-                            Template.CatalogResultPageLazyLoaderTemplate?.let {
-                                it()
-                            }
-                        } else {
-                            Row(
-                                modifier = FavoritePageStyle.loadMoreContainer,
-                                Arrangement.Center
-                            ) {
-                                CircularProgressIndicator(
-                                    color = FavoritePageColor.loaderColor,
-                                    modifier = FavoritePageStyle.loadMoreModifier
-                                )
-                            }
+                }
+            }
+            item(span = { GridItemSpan(columns) }) {
+                if (recipePageVM.currentState.isFetchingNewPage) {
+                    if (Template.CatalogResultPageLazyLoaderTemplate != null) {
+                        Template.CatalogResultPageLazyLoaderTemplate?.let {
+                            it()
+                        }
+                    } else {
+                        Row(
+                            modifier = FavoritePageStyle.loadMoreContainer,
+                            Arrangement.Center
+                        ) {
+                            CircularProgressIndicator(
+                                color = FavoritePageColor.loaderColor,
+                                modifier = FavoritePageStyle.loadMoreModifier
+                            )
                         }
                     }
                 }