+++ changelog.md
@@ -1,4 +1,9 @@
-## 3.1.1
+## 3.1.0
+- [FIX][AND] Fix Recipe card CTA text is now customizable
+- [FIX][AND] Fix modal now fit correctly in screen
+- [FIX][COR] Analytics domain, user agent and multiple sent show event
+- [FIX][IOS] Add templates and some public modifiers
+- [FEA][AND] Change lazy column into lazy grid in recipes list
 - [BRK][FIX][AND] Fix recipes page lazy loading by upgrading compose (1.3.1) and android SDK (33)
 
 ## 3.0.1